requirements.txt
flask==2.3.2
opencv-python==4.5.5.64
deepface==0.0.79
numpy==1.24.3
pydub==0.25.1
simpleaudio==1.0.0
speechrecognition==3.8.1
gtts==2.2.4
python-dotenv==1.0.0
google-generativeai==0.3.2
google-cloud-firestore==2.11.0
google-cloud-storage==2.7.0
google-cloud-speech==2.16.2
google-cloud-texttospeech==2.13.1